backend/requirements.txt
Django>=4.2,<5
djangorestframework
django-cors-headers
Pillow>=10.0.0
PyMySQL>=1.1.0
python-dotenv>=1.0.0
